+++ .github/workflows/build.yaml
@@ -1,23 +1,24 @@
 name: build
 on: [push, pull_request]
 jobs:
-
+  
   build:
     name: Build
     runs-on: ubuntu-latest
     steps:
-
-    - name: Set up Go 1.13
-      uses: actions/setup-go@v1
+    - uses: actions/checkout@v2
+    - name: Set up JDK 1.8
+      uses: actions/setup-java@v1
       with:
-        go-version: 1.13
-      id: go
+        java-version: 1.8
 
-    - name: Check out code into the Go module directory
-      uses: actions/checkout@v1
+    - name: Cache the Maven packages to speed up build
+      uses: actions/cache@v2
+      with:
+        path: ~/.m2/repository
+        key: ${{ runner.os }}-maven-${{ hashFiles('**/pom.xml') }}
+        restore-keys: ${{ runner.os }}-maven-
 
-    - name: Get dependencies
-      run: |
-        go get -v -t -d ./...
-    - name: Build
-      run: go build -v ./bin/nb
+    - name: Build with Maven
+      run: cd util/LdbcGoStep && mvn -B package
+
