+++ util/LdbcGoStep/pom.xml
@@ -2,7 +2,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.vesoft</groupId>
-  <artifactId>ldbc_go_step</artifactId>
+  <artifactId>LdbcGoStep</artifactId>
   <packaging>jar</packaging>
   <version>2</version>
   <name>Maven Quick Start Archetype</name>
